requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.6.0
pydantic==2.9.2
httpx==0.27.2
python-dateutil==2.9.0.post0
tzdata>=2024.1
python-jose[cryptography]==3.3.0

cryptography==42.0.0
pywebpush==2.3.0
python-barcode==0.15.1
qrcode[pil]==7.4.2
Pillow>=11.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx>=0.27.2
respx==0.21.1
